Fix code to add validation function and excetopHandler

+++ APIDeliveryService/src/main/java/com/example/apideliveryservice/dto/CompanyFoodDto.java
@@ -1,7 +1,6 @@
 package com.example.apideliveryservice.dto;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,12 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CompanyFoodDto {
 
-    //bigint
-    private BigInteger id;
-    //bigint
-    private BigInteger memberId;
-    //varchar
+    private Long id;
+    private Long memberId;
     private String name;
-    //decimal
     private BigDecimal price;
 }
